WIP: refactor forms
